requirements.txt
Flask>=2.2
psycopg2-binary>=2.9
plotly>=5.0
numpy>=1.24
python-dotenv>=1.0
Pillow>=9.0
matplotlib>=3.7
